+++ linphone-app/src/utils/Utils.hpp
@@ -71,6 +71,7 @@ public:
 	Q_INVOKABLE static bool isVideo(const QString& path);
 	Q_INVOKABLE static bool isPdf(const QString& path);
 	Q_INVOKABLE static bool isSupportedForDisplay(const QString& path);
+	Q_INVOKABLE static bool canHaveThumbnail(const QString& path);
 	Q_INVOKABLE static bool isPhoneNumber(const QString& txt);
 	Q_INVOKABLE static bool isUsername(const QString& txt);	// Check with Regex
 	Q_INVOKABLE QSize getImageSize(const QString& url);