+++ linphone-app/ui/modules/Linphone/Styles/Chat/ChatStyle.qml
@@ -201,7 +201,17 @@ QtObject {
 					property var outgoingColor: ColorsList.addImageColor(sectionName+'_download_out', icon, 'g')
 					property var incomingColor: ColorsList.addImageColor(sectionName+'_download_in', icon, 'q')
 				}
-				
+				property QtObject thumbnailVideoIcon: QtObject {
+					property int iconSize: 40
+					property string name : 'play'
+					property string icon : 'thumbnail_video_custom'
+					property var backgroundNormalColor : ColorsList.addImageColor(sectionName+'_'+name+'_bg_n', icon, 'wr_n_b_bg')
+					property var backgroundHoveredColor : ColorsList.addImageColor(sectionName+'_'+name+'_bg_h', icon, 'wr_h_b_bg')
+					property var backgroundPressedColor : ColorsList.addImageColor(sectionName+'_'+name+'_bg_p', icon, 'wr_p_b_bg')
+					property var foregroundNormalColor : ColorsList.addImageColor(sectionName+'_'+name+'_fg_n', icon, 'wr_n_b_fg')
+					property var foregroundHoveredColor : ColorsList.addImageColor(sectionName+'_'+name+'_fg_h', icon, 'wr_h_b_fg')
+					property var foregroundPressedColor : ColorsList.addImageColor(sectionName+'_'+name+'_fg_p', icon, 'wr_p_b_fg')
+				}
 				property QtObject animation: QtObject {
 					property int duration: 300
 					property real to: 1.7
@@ -212,6 +222,7 @@ QtObject {
 					property string icon:  'file_extension_custom'
 					property string unknownIcon:  'file_unknown_custom'
 					property int iconSize: 60
+					property int internalSize: 37
 					property int radius: 0
 					
 					property QtObject background: QtObject {